+++ src/consts.js
@@ -1,33 +1,8 @@
-const POSTERS = [
-  'made-for-each-other.png',
-  'popeye-meets-sinbad.png',
-  'sagebrush-trail.jpg',
-  'santa-claus-conquers-the-martians.jpg',
-  'the-dance-of-life.jpg',
-  'the-great-flamarion.jpg',
-  'the-man-with-the-golden-arm.jpg'
-];
-
-const GENRES = [
-  'comedy',
-  'Musical',
-  'Western',
-  'drama',
-  'Cartoon'
-];
-
-const EMOTIONS = [
-  'smile',
-  'sleeping',
-  'puke',
-  'angry'
-];
-
 const DEFAULT_RENDERED_FILMS_QUANTITY = 5;
 const FILMS_TO_RENDER_QUANTITY = 5;
 const FILM_CARD_DATE_FORMAT = 'YYYY';
 const FILM_POPUP_DATE_FORMAT = 'D MMM YYYY';
 const FILM_COMMENT_DATE_FORMAT = 'YYYY/D/MM h:mm';
 
-export { POSTERS, GENRES, EMOTIONS, DEFAULT_RENDERED_FILMS_QUANTITY, FILMS_TO_RENDER_QUANTITY, FILM_CARD_DATE_FORMAT, FILM_POPUP_DATE_FORMAT, FILM_COMMENT_DATE_FORMAT };
+export { DEFAULT_RENDERED_FILMS_QUANTITY, FILMS_TO_RENDER_QUANTITY, FILM_CARD_DATE_FORMAT, FILM_POPUP_DATE_FORMAT, FILM_COMMENT_DATE_FORMAT };
