+++ src/mock/film.js
@@ -1,5 +1,29 @@
 import { getRandomArrayElement } from '../utils.js';
-import { POSTERS, GENRES, EMOTIONS } from '../consts.js';
+
+const POSTERS = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'santa-claus-conquers-the-martians.jpg',
+  'the-dance-of-life.jpg',
+  'the-great-flamarion.jpg',
+  'the-man-with-the-golden-arm.jpg'
+];
+
+const GENRES = [
+  'comedy',
+  'Musical',
+  'Western',
+  'drama',
+  'Cartoon'
+];
+
+const EMOTIONS = [
+  'smile',
+  'sleeping',
+  'puke',
+  'angry'
+];
 
 const mockComments = [
   {
